requirements.txt
gunicorn
python-dotenv
requests
pytest 
Flask
supabase
ollama
beautifulsoup4
duckduckgo-search
dirtyjson